server/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-rest-framework==0.1.0
python-jose==3.3.0
bcrypt==4.1.2
redis==5.0.3
django-redis==5.4.0
